chore: import latest server repo

+++ apps/server/src/workspaces/workspaces.service.ts
@@ -1,6 +1,6 @@
 import { ValidationError } from '@mikro-orm/core';
 import { InjectRepository } from '@mikro-orm/nestjs';
-import { Injectable } from '@nestjs/common';
+import { Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { Tenant } from 'src/tenants/entities/tenant.entity';
 import { TenantsService } from 'src/tenants/tenants.service';
 import { CreateWorkspaceInput } from './dto/create-workspace.input';
@@ -15,8 +15,9 @@ export class WorkspacesService {
     private readonly workspaceRepository: WorkspaceRepository,
     private readonly tenantService: TenantsService,
   ) {}
+  private readonly logger = new Logger('WorkspacesService');
 
-  async create(createWorkspaceInput: CreateWorkspaceInput) {
+  async create(createWorkspaceInput: CreateWorkspaceInput, user) {
     const tenant = await this.tenantService.findOne(
       createWorkspaceInput.tenantId,
     );
@@ -24,6 +25,8 @@ export class WorkspacesService {
       ...createWorkspaceInput,
       tenant,
     });
+    workspace.createdBy = user;
+    workspace.updatedBy = user;
     await this.workspaceRepository.persistAndFlush(workspace);
     return workspace;
   }
@@ -40,25 +43,50 @@ export class WorkspacesService {
     return workspace;
   }
 
-  findAll() {
-    return this.workspaceRepository.findAll();
+  async findAll() {
+    const workspaces = await this.workspaceRepository.findAll();
+    if (!workspaces) {
+      throw new Error('No such workspace collection  found');
+    }
+    return workspaces;
   }
 
-  findOne(id: string) {
-    return this.workspaceRepository.findOne(id);
+  async findOne(id: string) {
+    const workspace = await this.workspaceRepository.findOne(id);
+    if (!workspace) {
+      throw new Error('No such workspace');
+    }
+    return workspace;
   }
 
-  update(id: string, updateWorkspaceInput: UpdateWorkspaceInput) {
-    const workspace = this.workspaceRepository.findOne(id);
-
-    // todo: update workspace
-
-    return this.workspaceRepository.persistAndFlush(workspace);
+  async update(updateWorkspaceInput: UpdateWorkspaceInput): Promise<Workspace> {
+    const workspace = await this.workspaceRepository.findOne(
+      updateWorkspaceInput.id,
+    );
+    if (!workspace) {
+      // console.error('Console message - No such workspace');
+      throw new NotFoundException('GraphQL Error - No such workspace');
+    }
+    workspace.name = updateWorkspaceInput.name;
+    workspace.description = updateWorkspaceInput.description;
+    try {
+      await this.workspaceRepository.persistAndFlush(workspace);
+    } catch (error) {
+      throw new NotFoundException('No such workspace', error);
+    }
+    return workspace;
   }
 
-  remove(id: string) {
-    // todo: last workspace cannot be deleted
-    const workspace = this.workspaceRepository.findOne(id);
-    return this.workspaceRepository.removeAndFlush(workspace);
+  async remove(id: string) {
+    const workspace = await this.workspaceRepository.findOne(id);
+    if (!workspace) {
+      throw new NotFoundException('GrapphQl error - No such workspace found');
+    }
+    try {
+      await this.workspaceRepository.removeAndFlush(workspace);
+    } catch (error) {
+      throw new NotFoundException('DB Error', error);
+    }
+    return workspace;
   }
 }
